+++ scripts/deck.js
@@ -1,0 +1,15 @@
+import { Card } from './card.js';
+
+export function Deck() {
+    this.cards = [];
+    const numbers = [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14];
+    const suits = ['clubs','diamonds','hearts','spades'];
+
+    suits.forEach(suit => numbers
+        .forEach(num => this.cards
+        .push(new Card({val: num, suit: suit}))));
+}
+
+Deck.prototype.shuffle = function() {
+    this.cards.sort(() => Math.random() - 0.5);
+}
