+++ styles/main.css
@@ -1,8 +1,8 @@
 :root {
     --button-color: rgb(188, 89, 213);
     --button-shadow: rgb(103, 28, 121);
-    --win-effect-time: 1.2s;
-    --war-effect-time: 0.4s;
+    --win-effect-time: 0.5s;
+    --war-effect-time: 0.333s;
 }
 
 body {
@@ -192,7 +192,7 @@ img {
 }
 
 .create-deck {
-    animation: fade-in 0.5s;
+    animation: fade-in 0.7s;
 }
 
 .face-up .card-inner {
@@ -233,7 +233,7 @@ img {
         z-index: 100;
     }
     40% {
-        transform: scale(1.4);
+        transform: scale(1.1);
         z-index: 100;
         border-radius: 10px;
         box-shadow: 0 0px 20px 10px rgb(255, 245, 64);
@@ -262,7 +262,7 @@ img {
         transform: translateX(0px);
     }
     40% {
-        transform: translateX(20px);
+        transform: translateX(15px);
     }
     100% {
         transform: translateX(0px);
@@ -274,7 +274,7 @@ img {
         transform: translateX(0px);
     }
     40% {
-        transform: translateX(-20px);
+        transform: translateX(-15px);
     }
     100% {
         transform: translateX(0px);
